SSCS-2598 Refactoring
Organise models classes into more meaningful packages
