+++ src/main/java/uk/gov/hmcts/reform/sscs/models/serialize/ccd/CaseData.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.sscs.models;
+package uk.gov.hmcts.reform.sscs.models.serialize.ccd;
 
 import lombok.Builder;
 import lombok.Value;
