+++ src/main/java/uk/gov/hmcts/reform/sscs/services/mapper/TransformJsonCasesToCaseData.java
@@ -3,14 +3,14 @@ package uk.gov.hmcts.reform.sscs.services.mapper;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.springframework.http.converter.json.Jackson2ObjectMapperBuilder;
 import org.springframework.stereotype.Service;
-import uk.gov.hmcts.reform.sscs.models.Appeal;
-import uk.gov.hmcts.reform.sscs.models.Appellant;
-import uk.gov.hmcts.reform.sscs.models.CaseData;
-import uk.gov.hmcts.reform.sscs.models.Identity;
-import uk.gov.hmcts.reform.sscs.models.Name;
-import uk.gov.hmcts.reform.sscs.models.gaps2.AppealCase;
-import uk.gov.hmcts.reform.sscs.models.gaps2.Gaps2Extract;
-import uk.gov.hmcts.reform.sscs.models.gaps2.Parties;
+import uk.gov.hmcts.reform.sscs.models.deserialize.gaps2.AppealCase;
+import uk.gov.hmcts.reform.sscs.models.deserialize.gaps2.Gaps2Extract;
+import uk.gov.hmcts.reform.sscs.models.deserialize.gaps2.Parties;
+import uk.gov.hmcts.reform.sscs.models.serialize.ccd.Appeal;
+import uk.gov.hmcts.reform.sscs.models.serialize.ccd.Appellant;
+import uk.gov.hmcts.reform.sscs.models.serialize.ccd.CaseData;
+import uk.gov.hmcts.reform.sscs.models.serialize.ccd.Identity;
+import uk.gov.hmcts.reform.sscs.models.serialize.ccd.Name;
 
 import java.io.IOException;
